AssimpImporter: rewrite using MeshData.
Thanks to the (almost) 100% code coverage this was a matter of 20
minutes. Previous tests using the original MeshData3D all pass, adapting
them is for the next commit to ensure we don't break any existing uses.

What is apparently not tested are the various error states -- changing
the message from mesh3D() to mesh() didn't cause any test failures. Hmm.

Compared to StanfordImporter on the same file the "speedup" was from 7.9
seconds to 7.2 seconds. That's extremely bad, but the plugin code is as
optimal as it can get. Blame Assimp, not Magnum.